+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## [2.4.0] - 
+## [2.4.0] - 2/7/2019
 ### Added
 - Add event entity for ToteDetails projection
 - Add EventTypes: "AggregateFixture"